myresyapi/views.py
from shopppCat.models import Product, Order
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework.request import Request
from rest_framework.views import APIView
from .serializers import ProductSerializer, OrderSerializer

class ProductListView(APIView):
    def get(self, request: Request) -> Response:
        products = Product.objects.all()
        serializer = ProductSerializer(products, many=True)
        return Response({'products': serializer.data})


class OrderListView(APIView):
    def get(self, request: Request) -> Response:
        orders = Order.objects.all()
        serializer = OrderSerializer(orders, many=True)
        return Response({'orders': serializer.data})


