requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.2
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
